requirements.txt
tensorflow==2.10.0
flask==2.2.3
pillow==9.5.0
numpy==1.24.3
gunicorn==20.1.0